add license
This code is copied from eloipool-scrypt which is a fork of
LukeJr's eloipool including its AGPLv3 license, but also includes
the original scrypt code under a different license. It is the
easiest to just keep AGPLv3 for this repo for everything except
the scrypt sources - that is already taken care of.

As the Python integration code in setup.py and scryptmodule.c do
not have their own copyright headers, this means copyright is
implicitly assigned to the FSF. This commit makes that explicit.